Rename SearchDocument to SearchableDocument
